Extract feather utilities into dedicated modules
