+++ src/parser/src/transforms/apply-feather-fixes.ts
@@ -2,15 +2,14 @@
 
 import { Core, type MutableGameMakerAstNode } from "@gml-modules/core";
 import { FunctionalParserTransform } from "./functional-transform.js";
-import antlr4, { PredictionMode } from "antlr4";
-import GameMakerLanguageLexer from "../../generated/GameMakerLanguageLexer.js";
-import GameMakerLanguageParser from "../../generated/GameMakerLanguageParser.js";
-import GameMakerASTBuilder from "../ast/gml-ast-builder.js";
-import type { ParserContextWithMethods } from "../types/index.js";
-import { defaultParserOptions } from "../types/parser-types.js";
-import GameMakerParseErrorListener, {
-    GameMakerLexerErrorListener
-} from "../ast/gml-syntax-error.js";
+import {
+    getEndFromNode,
+    getStartFromNode,
+    hasArrayParentWithNumericIndex,
+    resolveCallExpressionArrayContext,
+    walkAstNodes
+} from "./feather/ast-traversal.js";
+import { parseExample } from "./feather/parser-bootstrap.js";
 import { preprocessFunctionArgumentDefaults } from "./preprocess-function-argument-defaults.js";
 
 type RenameOptions = {
@@ -27,121 +26,6 @@ type ApplyFeatherFixesOptions = {
     preprocessedFixMetadata?: unknown;
     options?: Record<string, unknown>;
 };
-
-function walkAstNodes(root, visitor) {
-    // TODO: I think this is duplicated elsewhere
-    const visit = (node, parent, key) => {
-        if (!node) {
-            return;
-        }
-
-        if (Array.isArray(node)) {
-            for (let index = 0; index < node.length; index += 1) {
-                visit(node[index], node, index);
-            }
-
-            return;
-        }
-
-        if (!Core.isNode(node)) {
-            return;
-        }
-
-        const shouldDescend = visitor(node, parent, key);
-
-        if (shouldDescend === false) {
-            return;
-        }
-
-        for (const [childKey, childValue] of Object.entries(node)) {
-            if (childValue && typeof childValue === "object") {
-                visit(childValue, node, childKey);
-            }
-        }
-    };
-
-    visit(root, null, null);
-}
-
-function hasArrayParentWithNumericIndex(parent, property) {
-    if (!Array.isArray(parent)) {
-        return false;
-    }
-
-    if (typeof property !== "number") {
-        return false;
-    }
-
-    return true;
-}
-
-function resolveCallExpressionArrayContext(node, parent, property) {
-    if (!hasArrayParentWithNumericIndex(parent, property)) {
-        return null;
-    }
-
-    if (!Core.isNode(node) || node.type !== "CallExpression") {
-        return null;
-    }
-
-    return {
-        callExpression: node,
-        siblings: parent,
-        index: property
-    };
-}
-
-function getStartFromNode(node: unknown) {
-    if (!Core.isNode(node)) return null;
-    if (!Object.hasOwn(node, "start")) return null;
-    return Core.cloneLocation((node as any).start);
-}
-
-function getEndFromNode(node: unknown) {
-    if (!Core.isNode(node)) return null;
-    if (!Object.hasOwn(node, "end")) return null;
-    return Core.cloneLocation((node as any).end);
-}
-
-function parseExample(
-    sourceText: string,
-    options: { getLocations?: boolean; simplifyLocations?: boolean } = {
-        getLocations: true,
-        simplifyLocations: false
-    }
-) {
-    if (typeof sourceText !== "string" || sourceText.length === 0) {
-        return null;
-    }
-
-    try {
-        const chars = new antlr4.InputStream(sourceText);
-        const lexer = new GameMakerLanguageLexer(chars);
-        lexer.removeErrorListeners();
-        lexer.addErrorListener(new GameMakerLexerErrorListener());
-        lexer.strictMode = false;
-        const tokens = new antlr4.CommonTokenStream(lexer);
-        const parser = new GameMakerLanguageParser(tokens);
-
-        parser._interp.predictionMode = PredictionMode.SLL;
-        parser.removeErrorListeners();
-        parser.addErrorListener(new GameMakerParseErrorListener());
-
-        const tree = parser.program();
-        const builder = new GameMakerASTBuilder(
-            {
-                ...defaultParserOptions,
-                getLocations: options.getLocations ?? true,
-                simplifyLocations: options.simplifyLocations ?? false
-            },
-            []
-        );
-        return builder.build(tree as ParserContextWithMethods);
-    } catch {
-        // Parsing example failed — return null and let caller handle absence
-        return null;
-    }
-}
 
 const TRAILING_MACRO_SEMICOLON_PATTERN = new RegExp(
     ";(?=[^\\S\\r\\n]*(?:(?:\\/\\/[^\\r\\n]*|\\/\\*[\\s\\S]*?\\*\/)[^\\S\\r\\n]*)*(?:\\r?\\n|$))"
@@ -306,517 +190,6 @@ const FEATHER_DIAGNOSTIC_FIXERS = buildFeatherDiagnosticFixers(
     FEATHER_DIAGNOSTICS,
     FEATHER_FIX_IMPLEMENTATIONS
 );
-
-export function preprocessSourceForFeatherFixes(sourceText) {
-    if (!Core.isNonEmptyString(sourceText)) {
-        return {
-            sourceText,
-            metadata: null
-        };
-    }
-
-    const gm1100Metadata = [];
-    const gm1016Metadata = [];
-    const sanitizedParts = [];
-    const newlinePattern = /\r?\n/g;
-    let lastIndex = 0;
-    let lineNumber = 1;
-    let pendingGM1100Context = null;
-
-    const processLine = (line) => {
-        const indentationMatch = line.match(/^\s*/);
-        const indentation = indentationMatch ? indentationMatch[0] : "";
-        const trimmed = Core.toTrimmedString(line);
-
-        if (trimmed.length === 0) {
-            return { line, context: pendingGM1100Context };
-        }
-
-        const booleanLiteralMatch = line.match(
-            /^(\s*)(true|false)\s*(?:;\s*)?$/
-        );
-
-        if (booleanLiteralMatch) {
-            const leadingWhitespace = booleanLiteralMatch[1] ?? "";
-            const sanitizedRemainder = " ".repeat(
-                Math.max(0, line.length - leadingWhitespace.length)
-            );
-            const sanitizedLine = `${leadingWhitespace}${sanitizedRemainder}`;
-            const trimmedRightLength = line.replace(/\s+$/, "").length;
-            const startColumn = leadingWhitespace.length;
-            const endColumn = Math.max(startColumn, trimmedRightLength - 1);
-            const lineStartIndex = lastIndex;
-
-            gm1016Metadata.push({
-                start: {
-                    line: lineNumber,
-                    column: startColumn,
-                    index: lineStartIndex + startColumn
-                },
-                end: {
-                    column: endColumn,
-                    index: lineStartIndex + endColumn
-                }
-            });
-
-            return { line: sanitizedLine, context: null };
-        }
-
-        const varMatch = line.match(/^\s*var\s+([A-Za-z_][A-Za-z0-9_]*)\b/);
-
-        if (varMatch) {
-            const identifier = varMatch[1];
-            const remainder = line.slice(varMatch[0].length);
-            const trimmedRemainder = remainder.replace(/^\s*/, "");
-
-            if (trimmedRemainder.startsWith("*")) {
-                const leadingWhitespaceLength =
-                    remainder.length - trimmedRemainder.length;
-                const leadingWhitespace =
-                    leadingWhitespaceLength > 0
-                        ? remainder.slice(0, leadingWhitespaceLength)
-                        : "";
-                const sanitizedLine = [
-                    line.slice(0, varMatch[0].length),
-                    leadingWhitespace,
-                    "=",
-                    trimmedRemainder.slice(1)
-                ].join("");
-
-                gm1100Metadata.push({
-                    type: "declaration",
-                    line: lineNumber,
-                    identifier
-                });
-
-                return {
-                    line: sanitizedLine,
-                    context: {
-                        identifier,
-                        indentation
-                    }
-                };
-            }
-        }
-
-        if (trimmed.startsWith("=") && pendingGM1100Context?.identifier) {
-            const rawRemainder = line.slice(indentation.length);
-            const identifier = pendingGM1100Context.identifier;
-
-            gm1100Metadata.push({
-                type: "assignment",
-                line: lineNumber,
-                identifier
-            });
-
-            const sanitizedLine = `${indentation}${" ".repeat(
-                Math.max(0, rawRemainder.length)
-            )}`;
-
-            return { line: sanitizedLine, context: null };
-        }
-
-        if (trimmed.startsWith("/") || trimmed.startsWith("*")) {
-            return { line, context: pendingGM1100Context };
-        }
-
-        return { line, context: null };
-    };
-
-    let match;
-
-    while ((match = newlinePattern.exec(sourceText)) !== null) {
-        const lineEnd = match.index;
-        const line = sourceText.slice(lastIndex, lineEnd);
-        const newline = match[0];
-        const { line: sanitizedLine, context } = processLine(line);
-
-        sanitizedParts.push(sanitizedLine, newline);
-        pendingGM1100Context = context;
-        lastIndex = match.index + newline.length;
-        lineNumber += 1;
-    }
-
-    const finalLine = sourceText.slice(lastIndex);
-    if (
-        finalLine.length > 0 ||
-        sourceText.endsWith("\n") ||
-        sourceText.endsWith("\r")
-    ) {
-        const { line: sanitizedLine, context } = processLine(finalLine);
-        sanitizedParts.push(sanitizedLine);
-        pendingGM1100Context = context;
-    }
-
-    const sanitizedSourceText = sanitizedParts.join("");
-    const enumSanitizedResult =
-        sanitizeEnumInitializerStrings(sanitizedSourceText);
-    const enumSanitizedSourceText = enumSanitizedResult.sourceText;
-    const enumIndexAdjustments = enumSanitizedResult.adjustments;
-    const metadata: any = {};
-
-    if (gm1100Metadata.length > 0) {
-        metadata.GM1100 = gm1100Metadata;
-    }
-
-    if (gm1016Metadata.length > 0) {
-        metadata.GM1016 = gm1016Metadata;
-    }
-
-    const hasMetadata = Object.keys(metadata).length > 0;
-    const sourceChanged = enumSanitizedSourceText !== sourceText;
-    const hasIndexAdjustments = Core.isNonEmptyArray(enumIndexAdjustments);
-
-    if (!hasMetadata && !sourceChanged) {
-        return {
-            sourceText,
-            metadata: null,
-            indexAdjustments: null
-        };
-    }
-
-    return {
-        sourceText: sourceChanged ? enumSanitizedSourceText : sourceText,
-        metadata: hasMetadata ? metadata : null,
-        indexAdjustments: hasIndexAdjustments ? enumIndexAdjustments : null
-    };
-}
-
-function sanitizeEnumInitializerStrings(sourceText) {
-    if (!Core.isNonEmptyString(sourceText)) {
-        return { sourceText, adjustments: null };
-    }
-
-    const enumPattern = /\benum\b/g;
-    let lastIndex = 0;
-    let result = "";
-    const adjustments = [];
-    let totalRemoved = 0;
-
-    while (enumPattern.exec(sourceText) !== null) {
-        const openBraceIndex = findNextOpenBrace(
-            sourceText,
-            enumPattern.lastIndex
-        );
-        if (openBraceIndex === -1) {
-            break;
-        }
-
-        const closeBraceIndex = findMatchingClosingBrace(
-            sourceText,
-            openBraceIndex
-        );
-
-        if (closeBraceIndex === -1) {
-            break;
-        }
-
-        result += sourceText.slice(lastIndex, openBraceIndex + 1);
-
-        const bodyStartIndex = openBraceIndex + 1;
-        const body = sourceText.slice(bodyStartIndex, closeBraceIndex);
-        const {
-            sanitizedBody,
-            adjustments: bodyAdjustments,
-            removedCount: bodyRemoved
-        } = sanitizeEnumBodyInitializerStrings(
-            body,
-            bodyStartIndex,
-            totalRemoved
-        );
-
-        if (bodyAdjustments.length > 0) {
-            adjustments.push(...bodyAdjustments);
-            totalRemoved += bodyRemoved;
-        }
-
-        result += sanitizedBody;
-        lastIndex = closeBraceIndex;
-        enumPattern.lastIndex = closeBraceIndex;
-    }
-
-    if (lastIndex === 0) {
-        return { sourceText, adjustments: null };
-    }
-
-    result += sourceText.slice(lastIndex);
-    return {
-        sourceText: result,
-        adjustments: adjustments.length > 0 ? adjustments : null
-    };
-}
-
-function sanitizeEnumBodyInitializerStrings(
-    body,
-    bodyStartIndex,
-    totalRemoved
-) {
-    if (!Core.isNonEmptyString(body)) {
-        return { sanitizedBody: body, adjustments: [], removedCount: 0 };
-    }
-
-    let bodyRemoved = 0;
-    const adjustments = [];
-
-    const sanitizedBody = body.replaceAll(
-        /(\s*[A-Za-z_][A-Za-z0-9_]*\s*=\s*)(["'])([^"']*)(\2)/g,
-        (fullMatch, prefix, _quote, rawValue, _closingQuote, offset) => {
-            const normalizedValue = rawValue.trim();
-            if (!isIntegerLiteralString(normalizedValue)) {
-                return fullMatch;
-            }
-
-            const replacement = `${prefix}${normalizedValue}`;
-            const removedCount = fullMatch.length - replacement.length;
-
-            if (removedCount > 0) {
-                const sanitizedIndex =
-                    bodyStartIndex +
-                    offset +
-                    replacement.length -
-                    (totalRemoved + bodyRemoved);
-
-                adjustments.push({
-                    index: sanitizedIndex,
-                    delta: removedCount
-                });
-                bodyRemoved += removedCount;
-            }
-
-            return replacement;
-        }
-    );
-
-    return { sanitizedBody, adjustments, removedCount: bodyRemoved };
-}
-
-export function applyRemovedIndexAdjustments(target, adjustments) {
-    const normalized = normalizeRemovalAdjustments(adjustments);
-    if (normalized.length === 0) {
-        return;
-    }
-
-    const stack = [target];
-    const seen = new WeakSet();
-
-    while (stack.length > 0) {
-        const current = stack.pop();
-
-        if (!current || typeof current !== "object" || seen.has(current)) {
-            continue;
-        }
-
-        seen.add(current);
-
-        if (Array.isArray(current)) {
-            for (const value of current) {
-                stack.push(value);
-            }
-            continue;
-        }
-
-        adjustLocationForRemoval(current, "start", normalized);
-        adjustLocationForRemoval(current, "end", normalized);
-
-        for (const value of Object.values(current)) {
-            stack.push(value);
-        }
-    }
-}
-
-function normalizeRemovalAdjustments(adjustments) {
-    if (!Array.isArray(adjustments)) {
-        return [];
-    }
-
-    return adjustments
-        .filter((entry) => {
-            if (!entry || typeof entry !== "object") {
-                return false;
-            }
-
-            const { index, delta } = entry;
-            return (
-                Number.isFinite(index) && Number.isFinite(delta) && delta > 0
-            );
-        })
-        .sort((a, b) => a.index - b.index);
-}
-
-function adjustLocationForRemoval(node, property, adjustments) {
-    if (!Object.hasOwn(node, property)) {
-        return;
-    }
-
-    const location = node[property];
-
-    if (typeof location === "number") {
-        node[property] = mapIndexForRemoval(location, adjustments);
-        return;
-    }
-
-    if (
-        location &&
-        typeof location === "object" &&
-        typeof location.index === "number"
-    ) {
-        location.index = mapIndexForRemoval(location.index, adjustments);
-    }
-}
-
-function mapIndexForRemoval(index, adjustments) {
-    if (!Number.isFinite(index)) {
-        return index;
-    }
-
-    let adjusted = index;
-
-    for (const { index: cutoff, delta } of adjustments) {
-        if (index >= cutoff) {
-            adjusted += delta;
-        } else {
-            break;
-        }
-    }
-
-    return adjusted;
-}
-
-function findNextOpenBrace(sourceText, startIndex) {
-    const length = sourceText.length;
-
-    for (let index = startIndex; index < length; index += 1) {
-        const char = sourceText[index];
-
-        if (char === '"' || char === "'") {
-            index = skipStringLiteral(sourceText, index);
-            continue;
-        }
-
-        if (
-            char === "@" &&
-            index + 1 < length &&
-            (sourceText[index + 1] === '"' || sourceText[index + 1] === "'")
-        ) {
-            index = skipStringLiteral(sourceText, index + 1);
-            continue;
-        }
-
-        if (char === "/" && index + 1 < length) {
-            const nextChar = sourceText[index + 1];
-            if (nextChar === "/") {
-                index = skipLineComment(sourceText, index + 2);
-                continue;
-            }
-            if (nextChar === "*") {
-                index = skipBlockComment(sourceText, index + 2);
-                continue;
-            }
-        }
-
-        if (char === "{") {
-            return index;
-        }
-    }
-
-    return -1;
-}
-
-function findMatchingClosingBrace(sourceText, openBraceIndex) {
-    const length = sourceText.length;
-    let depth = 0;
-
-    for (let index = openBraceIndex; index < length; index += 1) {
-        const char = sourceText[index];
-
-        if (char === '"' || char === "'") {
-            index = skipStringLiteral(sourceText, index);
-            continue;
-        }
-
-        if (
-            char === "@" &&
-            index + 1 < length &&
-            (sourceText[index + 1] === '"' || sourceText[index + 1] === "'")
-        ) {
-            index = skipStringLiteral(sourceText, index + 1);
-            continue;
-        }
-
-        if (char === "/" && index + 1 < length) {
-            const nextChar = sourceText[index + 1];
-            if (nextChar === "/") {
-                index = skipLineComment(sourceText, index + 2);
-                continue;
-            }
-            if (nextChar === "*") {
-                index = skipBlockComment(sourceText, index + 2);
-                continue;
-            }
-        }
-
-        if (char === "{") {
-            depth += 1;
-            continue;
-        }
-
-        if (char === "}") {
-            depth -= 1;
-            if (depth === 0) {
-                return index;
-            }
-        }
-    }
-
-    return -1;
-}
-
-function skipLineComment(sourceText, startIndex) {
-    const length = sourceText.length;
-
-    for (let index = startIndex; index < length; index += 1) {
-        const char = sourceText[index];
-        if (char === "\n" || char === "\r") {
-            return index - 1;
-        }
-    }
-
-    return length - 1;
-}
-
-function skipBlockComment(sourceText, startIndex) {
-    const length = sourceText.length;
-
-    for (let index = startIndex; index < length - 1; index += 1) {
-        if (sourceText[index] === "*" && sourceText[index + 1] === "/") {
-            return index + 1;
-        }
-    }
-
-    return length - 1;
-}
-
-function skipStringLiteral(sourceText, startIndex) {
-    const length = sourceText.length;
-    const quote = sourceText[startIndex];
-    let index = startIndex + 1;
-
-    while (index < length) {
-        const char = sourceText[index];
-        if (char === "\\") {
-            index += 2;
-            continue;
-        }
-
-        if (char === quote) {
-            return index;
-        }
-
-        index += 1;
-    }
-
-    return length - 1;
-}
 
 export function getFeatherDiagnosticFixers() {
     return new Map(FEATHER_DIAGNOSTIC_FIXERS);
